Refactor to use singledispatch
This is a proposal to refactor the code to use singledispatch.

Description:

So far, we have written a custom dispatch logic by walking the mro until
we find a suitable object. Now we use a singledispatch decorator instead
which under the hood does something very similar but is based on the
builtin singledispatch decorator.

Pros:

It is no longer necessary to litter the code with get_state_method and
get_instance_method calls, as that is resolved through the decorator.
Also, we can assume that the singledispatch decorator is battle tested
and thus more robust than our custom implementation. Finally, it seems
to be using some form of caching, so maybe it's also faster. Finally,
since we no longer need to resolve the methods, the local imports are
gone.

Cons:

Unfortunately, we could not rely completely on the built in
singledispatch decorator. The reason is the following:

Usually, we want to dispatch on the class of the object. However, when
we call get_instance, the object is *always* a dict, which invalidates
the dispatch. Therefore, we change the dispatcher to dispatch on the
instance, not the class. By default, we just use the class of the
instance being passed, i.e. we do exactly the same as in the original
implementation. However, if we encounter a state dict, we resolve the
actual class from the state dict first and then dispatch on that class.

Another con is that we have to decorate all the functions that we plan
on using. Probably, this could be solved similarly to the previous
implementation by dynamically importing the functions and decorating
them on the fly, I have not tried.