Add support for CV and random states
Add support for RandomState and Generator. Unfortunately, the numpy docs
were a bit sparse when it comes to how to save and load those, I think I
got it right but please check. Especially, do you think the new test is
rigorous enough?

Regarding cross validators, ideally we would have a something like
estimator_checks but for cross validators instead. I didn't find
anything (LMK if there is) so for now I just tested a couple of cross
validators but not all.

As you can see, the cross validators just use BaseEstimator_get_state
and BaseEstimator_get_instance under the hood. I assume that many of the
failing tests can also be fixed by falling back on those two functions
when we encounter an sklearn class which is not taken care of (if it's
not Cython). This makes me wonder if we could/should have something like
a module level fallback. Since we control the dispatcher, it should be
possible to implement this.

Finally, I fixed the bug where sparse matrices were stored on the
project root.