Add more metainfo to persisted data
Each state (including sub-states) now contains the version of the
corresponding package. For stdlib and builtins, the Python
version (major, minor, micro) is used.

Implementation

The metainfo has just been added to the schema without a special key
like "metainfo", as has been discussed previously.

Metainfo is stored on all sub-states, i.e. there is a lot of redundancy.
If there is a better approach, LMK.

In order to avoid having to touch all get_state functions to add the
package version, it is instead added generically by the dispatch wrapper
function.

One wart of this PR is that the list of stdlib modules is hard-coded.
There is a method to get them in Python 3.10 but not earlier. Thus, it
could be possible that the list is not exactly correct for other Python
versions. Moreover, it will probably lead to wrong results if the
package name is shadowed (say, someone pip installs a package whose name
is "collections").