Fix for persistince numpy arrays of object dtype
Previously, when a numpy array could not be stored directly, it was
attempted to json-serialize the object. However, this didn't work. The
new solution now recursively calls get_state on the objects of the
array.

To check the effect, the testing had to be fixed to be able to compare
generic objects (à la Foo() == Foo()). This in turn unmasked some
existing bugs where the dtype might change after a save-load
roundtrip ([OneHotEncoder()] - AssertionError: assert dtype('<U6') ==
object).

It also reveals that for SGDClassifier and SGDOneClassSVM, there
actually seem to be incorrectly stored values.

Locally, test failures went from 15 to 13.