Add support for BallTree, BinaryTree
Those are used by KNeighbors{Classifier,Regressor} but were not caught
so far because of the default argument.

Unfortunately, we need some special treatment here because the two
cython classes, as well as DistanceMetric, which is used under the hood,
return tuples from __getstate__ and thus the normal reduce_get_state and
reduce_get_instance methods fail.