Add support for partial
partial uses reduce/__getstate__/__setstate__ under the hood. To be on
the safe side, we hard-code the use of the partial class instance of
having a dynamically determined class. This should be okay
because (AFAIK) there are no partial subclasses in Python. If a user
provides a partial subclass, they have to provide their own dispatch
functions.